Move Ryzen related fixes to RxFix class.
