+++ src/crypto/rx/Rx.cpp
@@ -1,7 +1,7 @@
 /* XMRig
  * Copyright (c) 2018-2019 tevador     <tevador@gmail.com>
- * Copyright (c) 2018-2020 SChernykh   <https://github.com/SChernykh>
- * Copyright (c) 2016-2020 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
+ * Copyright (c) 2018-2021 SChernykh   <https://github.com/SChernykh>
+ * Copyright (c) 2016-2021 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
@@ -25,6 +25,11 @@
 #include "crypto/rx/RxQueue.h"
 #include "crypto/randomx/randomx.h"
 #include "crypto/randomx/aes_hash.hpp"
+
+
+#ifdef XMRIG_FIX_RYZEN
+#   include "crypto/rx/RxFix.h"
+#endif
 
 
 namespace xmrig {
@@ -104,7 +109,10 @@ bool xmrig::Rx::init(const T &seed, const RxConfig &config, const CpuConfig &cpu
     }
 
     if (!osInitialized) {
-        setupMainLoopExceptionFrame();
+#       ifdef XMRIG_FIX_RYZEN
+        RxFix::setupMainLoopExceptionFrame();
+#       endif
+
         if (!cpu.isHwAES()) {
             SelectSoftAESImpl(cpu.threads().get(seed.algorithm()).count());
         }
