+++ src/crypto/rx/RxFix.h
@@ -17,54 +17,26 @@
  *   along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef XMRIG_RX_H
-#define XMRIG_RX_H
+#ifndef XMRIG_RXFIX_H
+#define XMRIG_RXFIX_H
 
 
-#include <cstdint>
 #include <utility>
-#include <vector>
-
-
-#include "crypto/common/HugePagesInfo.h"
 
 
 namespace xmrig
 {
 
 
-class Algorithm;
-class CpuConfig;
-class CpuThread;
-class IRxListener;
-class Job;
-class RxConfig;
-class RxDataset;
-
-
-class Rx
+class RxFix
 {
 public:
-    static HugePagesInfo hugePages();
-    static RxDataset *dataset(const Job &job, uint32_t nodeId);
-    static void destroy();
-    static void init(IRxListener *listener);
-    template<typename T> static bool init(const T &seed, const RxConfig &config, const CpuConfig &cpu);
-    template<typename T> static bool isReady(const T &seed);
-
-#   ifdef XMRIG_FEATURE_MSR
-    static bool isMSR();
-#   else
-    static constexpr bool isMSR()   { return false; }
-#   endif
-
-private:
-    static bool msrInit(const RxConfig &config, const std::vector<CpuThread>& threads);
-    static void msrDestroy();
+    static void setMainLoopBounds(const std::pair<const void *, const void *> &bounds);
+    static void setupMainLoopExceptionFrame();
 };
 
 
 } /* namespace xmrig */
 
 
-#endif /* XMRIG_RX_H */
+#endif /* XMRIG_RXFIX_H */
