+++ src/crypto/rx/Rx_linux.cpp
@@ -37,8 +37,6 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <unistd.h>
-#include <signal.h>
-#include <ucontext.h>
 
 
 namespace xmrig {
@@ -235,33 +233,6 @@ static bool wrmsr(const MsrItems& preset, const std::vector<CpuThread>& threads,
 }
 
 
-#ifdef XMRIG_FIX_RYZEN
-static thread_local std::pair<const void*, const void*> mainLoopBounds = { nullptr, nullptr };
-
-static void MainLoopHandler(int sig, siginfo_t *info, void *ucontext)
-{
-    ucontext_t *ucp = (ucontext_t*) ucontext;
-
-    LOG_VERBOSE(YELLOW_BOLD("%s at %p"), (sig == SIGSEGV) ? "SIGSEGV" : "SIGILL", ucp->uc_mcontext.gregs[REG_RIP]);
-
-    void* p = reinterpret_cast<void*>(ucp->uc_mcontext.gregs[REG_RIP]);
-    const std::pair<const void*, const void*>& loopBounds = mainLoopBounds;
-
-    if ((loopBounds.first <= p) && (p < loopBounds.second)) {
-        ucp->uc_mcontext.gregs[REG_RIP] = reinterpret_cast<size_t>(loopBounds.second);
-    }
-    else {
-        abort();
-    }
-}
-
-void Rx::setMainLoopBounds(const std::pair<const void*, const void*>& bounds)
-{
-    mainLoopBounds = bounds;
-}
-#endif
-
-
 } // namespace xmrig
 
 
@@ -298,16 +269,4 @@ void xmrig::Rx::msrDestroy()
     if (!wrmsr(savedState, std::vector<CpuThread>(), true, false)) {
         LOG_ERR(CLEAR "%s" RED_BOLD_S "failed to restore initial state" BLACK_BOLD(" (%" PRIu64 " ms)"), tag, Chrono::steadyMSecs() - ts);
     }
-}
-
-
-void xmrig::Rx::setupMainLoopExceptionFrame()
-{
-#   ifdef XMRIG_FIX_RYZEN
-    struct sigaction act = {};
-    act.sa_sigaction = MainLoopHandler;
-    act.sa_flags = SA_RESTART | SA_SIGINFO;
-    sigaction(SIGSEGV, &act, nullptr);
-    sigaction(SIGILL, &act, nullptr);
-#   endif
 }
