+++ src/crypto/rx/Rx_win.cpp
@@ -348,43 +348,6 @@ static bool wrmsr(const MsrItems &preset, const std::vector<CpuThread>& threads,
 }
 
 
-#ifdef XMRIG_FIX_RYZEN
-static thread_local std::pair<const void*, const void*> mainLoopBounds = { nullptr, nullptr };
-
-static LONG WINAPI MainLoopHandler(_EXCEPTION_POINTERS *ExceptionInfo)
-{
-    if (ExceptionInfo->ExceptionRecord->ExceptionCode == 0xC0000005) {
-        const char* accessType;
-        switch (ExceptionInfo->ExceptionRecord->ExceptionInformation[0]) {
-        case 0: accessType = "read"; break;
-        case 1: accessType = "write"; break;
-        case 8: accessType = "DEP violation"; break;
-        default: accessType = "unknown"; break;
-        }
-        LOG_VERBOSE(YELLOW_BOLD("[THREAD %u] Access violation at 0x%p: %s at address 0x%p"), GetCurrentThreadId(), ExceptionInfo->ExceptionRecord->ExceptionAddress, accessType, ExceptionInfo->ExceptionRecord->ExceptionInformation[1]);
-    }
-    else {
-        LOG_VERBOSE(YELLOW_BOLD("[THREAD %u] Exception 0x%08X at 0x%p"), GetCurrentThreadId(), ExceptionInfo->ExceptionRecord->ExceptionCode, ExceptionInfo->ExceptionRecord->ExceptionAddress);
-    }
-
-    void* p = reinterpret_cast<void*>(ExceptionInfo->ContextRecord->Rip);
-    const std::pair<const void*, const void*>& loopBounds = mainLoopBounds;
-
-    if ((loopBounds.first <= p) && (p < loopBounds.second)) {
-        ExceptionInfo->ContextRecord->Rip = reinterpret_cast<DWORD64>(loopBounds.second);
-        return EXCEPTION_CONTINUE_EXECUTION;
-    }
-
-    return EXCEPTION_CONTINUE_SEARCH;
-}
-
-void Rx::setMainLoopBounds(const std::pair<const void*, const void*>& bounds)
-{
-    mainLoopBounds = bounds;
-}
-#endif
-
-
 } // namespace xmrig
 
 
@@ -420,12 +383,4 @@ void xmrig::Rx::msrDestroy()
     if (!wrmsr(savedState, std::vector<CpuThread>(), true, false)) {
         LOG_ERR(CLEAR "%s" RED_BOLD_S "failed to restore initial state" BLACK_BOLD(" (%" PRIu64 " ms)"), tag, Chrono::steadyMSecs() - ts);
     }
-}
-
-
-void xmrig::Rx::setupMainLoopExceptionFrame()
-{
-#   ifdef XMRIG_FIX_RYZEN
-    AddVectoredExceptionHandler(1, MainLoopHandler);
-#   endif
 }
